finished play music function

+++ ArchivioCD/src/archive/Song.java
@@ -1,11 +1,21 @@
 package archive;
 
+/**
+ * Class to manage the songs
+ * @author Andrea Bellu
+ */
 public class Song {
     String title;
-    int duration;
+    double duration;
     int year;
 
-    public Song(String title, int duration, int year) {
+    /**
+     * Constructor for the Song class
+     * @param title the title of the song
+     * @param duration the duration of the song
+     * @param year the year of the song
+     */
+    public Song(String title, double duration, int year) {
         this.title = title;
         this.duration = duration;
         this.year = year;
